demo-ahp/requirements.txt
flask
flask-cors
mysql-connector-python
 
#  pip install flask flask-cors mysql-connector-python